+++ tests/testthat/test_mlr_measures_surv.R
@@ -1,3 +1,4 @@
+# construct `pred` object with all possible predict types for testing
 set.seed(1L)
 task = tsk("rats")$filter(sample(300, 50L))
 learner = suppressWarnings(lrn("surv.coxph")$train(task))
@@ -29,20 +30,7 @@ test_that("mlr_measures", {
       perf = pred$score(m, task = task, train_set = task$row_ids, learner = learner)
     })
     expect_number(perf, na.ok = "na_score" %in% m$properties)
-
-    # test measures with squared-errors (se = TRUE)
-    if (key %in% paste0("surv.", c("schmid", "graf", "intlogloss", "logloss", "mae", "mse",
-      "rmse", "calib_alpha", "calib_beta"))) {
-      m = suppressWarnings(msr(key, se = TRUE))
-      perf = pred$score(m, task = task, train_set = task$row_ids, learner = learner)
-      expect_number(perf, na.ok = TRUE)
-    }
   }
-})
-
-test_that("unintegrated_prob_losses", {
-  msr = msr("surv.logloss")
-  expect_silent(pred$score(msr))
 })
 
 test_that("integrated losses with use of times", {
@@ -75,10 +63,86 @@ test_that("integrated losses with use of times", {
 })
 
 test_that("dcalib works", {
+  # see if chisq statistic and associated p-value are correctedly calculated
+  dc1 = msr("surv.dcalib", B = 14, chisq = FALSE) # default
+  dc2 = msr("surv.dcalib", B = 14, chisq = TRUE)
   expect_equal(
-    pchisq(pred$score(msr("surv.dcalib", B = 14)), df = 13, lower.tail = FALSE),
-    suppressWarnings(pred$score(msr("surv.dcalib", B = 14, chisq = TRUE)))
+    pchisq(pred$score(dc1), df = 13, lower.tail = FALSE),
+    suppressWarnings(pred$score(dc2))
   )
+
+  set.seed(1L)
+  t = tsk("rats")$filter(sample(1:300, 50))
+  l = lrn("surv.coxph")
+  p = suppressWarnings(l$train(t)$predict(t))
+
+  m = msr("surv.dcalib", truncate = 20)
+  expect_true(m$minimize)
+  expect_equal(m$range, c(0, Inf))
+  expect_equal(m$param_set$values$B, 10)
+  expect_false(m$param_set$values$chisq)
+  expect_equal(m$param_set$values$truncate, 20)
+  KMscore = p$score(m)
+  expect_numeric(KMscore)
+
+  status = t$truth()[, 2L]
+  row_ids = t$row_ids
+  cens_ids = row_ids[status == 0]
+  event_ids = row_ids[status == 1]
+
+  # only censored rats in test set
+  p = l$predict(t, row_ids = cens_ids)
+  score = p$score(m)
+  expect_numeric(score)
+
+  # 1 censored test rat
+  p = p$filter(row_ids = cens_ids[1L])
+  score = p$score(m)
+  expect_numeric(score)
+
+  # only dead rats in test set
+  p = l$predict(t, row_ids = event_ids)
+  score = p$score(m)
+  expect_numeric(score)
+
+  # 1 dead rat
+  p = p$filter(row_ids = event_ids[1L])
+  score = p$score(m)
+  expect_numeric(score)
+
+  # Another edge case: some dead rats and 1 only censored
+  p = l$predict(t, row_ids = c(event_ids, cens_ids[1L]))
+  score = p$score(m)
+  expect_numeric(score)
+  expect_gt(score, 10)
+
+  score3 = p$score(msr("surv.dcalib", truncate = 10))
+  expect_equal(unname(score3), 10)
+  score4 = p$score(msr("surv.dcalib", truncate = 5))
+  expect_equal(unname(score4), 5)
+  score5 = p$score(msr("surv.dcalib", truncate = Inf, B = 20)) # B affects truncate
+  expect_gt(score5, score)
+})
+
+test_that("mae/mse/rmse work", {
+  is_event = pred$truth[, 2L] == 1
+  event_times = pred$truth[is_event, 1L]
+  surv_times  = pred$response[is_event]
+  errors = event_times - surv_times
+  mae = pred$score(msr("surv.mae"))
+  mse = pred$score(msr("surv.mse"))
+  rmse = pred$score(msr("surv.rmse"))
+
+  expect_equal(unname(mae), mean(abs(errors)))
+  expect_equal(unname(mse), mean(errors^2))
+  expect_equal(unname(rmse), sqrt(mean(errors^2)))
+
+  # only censored observations => return NA
+  cens_ids = pred$row_ids[pred$truth[, 2L] == 0]
+  p_cens = pred$clone()$filter(cens_ids)
+  expect_true(is.na(p_cens$score(msr("surv.mae"))))
+  expect_true(is.na(p_cens$score(msr("surv.mse"))))
+  expect_true(is.na(p_cens$score(msr("surv.rmse"))))
 })
 
 test_that("calib_beta works", {
@@ -242,13 +306,15 @@ test_that("cindex: t_max, p_max", {
   expect_true(c1 != c3)
 })
 
-test_that("ERV works as expected", {
+test_that("ERV works", {
   set.seed(1L)
   t = tsk("rats")
   part = partition(t, 0.8)
   l = lrn("surv.kaplan")
   p = l$train(t, part$train)$predict(t, part$test)
   m = msr("surv.graf", ERV = TRUE)
+  # ERV = TRUE needs train task
+  expect_error(p$score(m))
   # KM is the baseline score, so ERV score = 0
   expect_equal(as.numeric(p$score(m, task = t, train_set = part$train)), 0)
 
@@ -257,15 +323,6 @@ test_that("ERV works as expected", {
   m = msr("surv.graf", ERV = TRUE)
   # Cox should do a little better than the KM baseline (ERV score > 0)
   expect_gt(as.numeric(p$score(m, task = t, train_set = part$train)), 0)
-
-  # some checks
-  set.seed(1L)
-  t = tsk("rats")$filter(sample(1:300, 50))
-  l = lrn("surv.kaplan")
-  p = l$train(t)$predict(t)
-  m = msr("surv.graf", ERV = TRUE, se = TRUE)
-  expect_error(p$score(m), "'task'")
-  expect_error(p$score(m, task = t, train_set = t$row_ids), "`se`")
 })
 
 test_that("ERV=TRUE changes some measure fields", {
@@ -320,15 +377,12 @@ test_that("rcll works", {
   t = tsk("rats")$filter(sample(1:300, 50))
   l = lrn("surv.kaplan")
   p = l$train(t)$predict(t)
-  p2 = p$clone()
-  p2$data$distr = p2$distr # hack: test score via distribution
+
   m = msr("surv.rcll")
   expect_true(m$minimize)
   expect_equal(m$range, c(0, Inf))
   KMscore = p$score(m)
   expect_numeric(KMscore)
-  KMscore2 = p2$score(m)
-  expect_equal(KMscore, KMscore2)
 
   status = t$truth()[, 2L]
   row_ids = t$row_ids
@@ -339,37 +393,21 @@ test_that("rcll works", {
   p = l$predict(t, row_ids = cens_ids)
   score = p$score(m)
   expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr
-  score2 = p2$score(m)
-  expect_equal(score, score2)
 
   # 1 censored test rat
   p = p$filter(row_ids = cens_ids[1L])
   score = p$score(m)
   expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr
-  score2 = p2$score(m)
-  expect_equal(score, score2)
 
   # only dead rats in test set
   p = l$predict(t, row_ids = event_ids)
   score = p$score(m)
   expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr # Matdist(1xY)
-  score2 = p2$score(m)
-  expect_equal(score, score2)
 
   # 1 dead rat
   p = p$filter(row_ids = event_ids[1L])
   score = p$score(m)
   expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr[1L] # WeightDisc
-  score2 = p2$score(m)
-  expect_equal(score, score2)
 
   # Cox is better than baseline (Kaplan-Meier)
   l2 = lrn("surv.coxph")
@@ -380,168 +418,31 @@ test_that("rcll works", {
   p3 = p2$filter(row_ids = c(event_ids, cens_ids[1L]))
   score = p3$score(m)
   expect_numeric(score)
-  p3$data$distr = p3$distr
-  score2 = p3$score(m)
-  expect_equal(score, score2)
-})
-
-test_that("dcal works", {
-  set.seed(1L)
-  t = tsk("rats")$filter(sample(1:300, 50))
-  l = lrn("surv.coxph")
-  p = suppressWarnings(l$train(t)$predict(t))
-  p2 = p$clone()
-  p2$data$distr = p2$distr # hack: test score via distribution
-  m = msr("surv.dcalib", truncate = 20)
-  expect_true(m$minimize)
-  expect_equal(m$range, c(0, Inf))
-  expect_equal(m$param_set$values$B, 10)
-  expect_false(m$param_set$values$chisq)
-  expect_equal(m$param_set$values$truncate, 20)
-  KMscore = p$score(m)
-  expect_numeric(KMscore)
-  KMscore2 = p2$score(m)
-  expect_equal(KMscore, KMscore2)
-
-  status = t$truth()[, 2L]
-  row_ids = t$row_ids
-  cens_ids = row_ids[status == 0]
-  event_ids = row_ids[status == 1]
-
-  # only censored rats in test set
-  p = l$predict(t, row_ids = cens_ids)
-  score = p$score(m)
-  expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr
-  score2 = p2$score(m)
-  expect_equal(score, score2)
-
-  # 1 censored test rat
-  p = p$filter(row_ids = cens_ids[1L])
-  score = p$score(m)
-  expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr
-  score2 = p2$score(m)
-  expect_equal(score, score2)
-
-  # only dead rats in test set
-  p = l$predict(t, row_ids = event_ids)
-  score = p$score(m)
-  expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr # Matdist(1xY)
-  score2 = p2$score(m)
-  expect_equal(score, score2)
-
-  # 1 dead rat
-  p = p$filter(row_ids = event_ids[1L])
-  score = p$score(m)
-  expect_numeric(score)
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr[1L] # WeightDisc
-  score2 = p2$score(m)
-  expect_equal(score, score2)
-
-  # Another edge case: some dead rats and 1 only censored
-  p = l$predict(t, row_ids = c(event_ids, cens_ids[1L]))
-  score = p$score(m)
-  expect_numeric(score)
-  p$data$distr = p$distr
-  score2 = p$score(m)
-  expect_equal(score, score2)
-  expect_gt(score, 10)
-
-  score3 = p$score(msr("surv.dcalib", truncate = 10))
-  expect_equal(unname(score3), 10)
-  score4 = p$score(msr("surv.dcalib", truncate = 5))
-  expect_equal(unname(score4), 5)
-  score5 = p$score(msr("surv.dcalib", truncate = Inf, B = 20)) # B affects truncate
-  expect_gt(score5, score)
 })
 
 test_that("logloss works", {
-  set.seed(1L)
-  t = tsk("rats")$filter(sample(1:300, 50))
-  l = lrn("surv.kaplan")
-  p = l$train(t)$predict(t)
-  p2 = p$clone()
-  p2$data$distr = p2$distr # hack: test score via distribution
-  m = msr("surv.logloss") # IPCW = TRUE (RNLL)
-  m2 = msr("surv.logloss", IPCW = FALSE) # NLL
+  t = tsk("lung")
+  kaplan = lrn("surv.kaplan")
+  cox = lrn("surv.coxph")
+  p_kaplan = kaplan$train(t)$predict(t)
+  p_cox = cox$train(t)$predict(t)
+
+  m = msr("surv.logloss")
   expect_true(m$minimize)
   expect_equal(m$range, c(0, Inf))
-  KMscore = p$score(m)
-  expect_numeric(KMscore)
-  KMscore2 = p2$score(m)
-  expect_equal(KMscore, KMscore2)
 
-  status = t$truth()[, 2L]
-  row_ids = t$row_ids
-  cens_ids = row_ids[status == 0]
-  event_ids = row_ids[status == 1]
+  kaplan_score = p_kaplan$score(m)
+  expect_numeric(kaplan_score)
+  cox_score = p_cox$score(m)
+  expect_numeric(cox_score)
 
-  # only censored rats in test set
-  p = l$predict(t, row_ids = cens_ids)
-  expect_true(is.nan(p$score(m))) # NaN
-  expect_false(is.nan(p$score(m2)))
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr
-  expect_true(is.nan(p2$score(m))) # NaN
-  expect_false(is.nan(p2$score(m2)))
+  # Cox is a bit better than baseline (Kaplan-Meier) in this dataset
+  expect_lt(cox_score, kaplan_score)
 
-  # 1 censored test rat
-  p = p$filter(row_ids = cens_ids[1L])
-  expect_true(is.nan(p$score(m)))
-  expect_false(is.nan(p$score(m2)))
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr
-  expect_true(is.nan(p2$score(m)))
-  expect_false(is.nan(p2$score(m2)))
-
-  # only dead rats in test set
-  p = l$predict(t, row_ids = event_ids)
-  score = p$score(m)
-  expect_numeric(score)
-  expect_equal(score, p$score(m2)) # as G_km(t) is 1
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr # Matdist(1xY)
-  score2 = p2$score(m)
-  expect_equal(score, score2)
-
-  # 1 dead rat
-  p = p$filter(row_ids = event_ids[1L])
-  score = p$score(m)
-  expect_numeric(score)
-  expect_equal(score, p$score(m2))
-  p2 = p$clone() # test score via distribution
-  p2$data$distr = p2$distr[1L] # WeightDisc
-  score2 = p2$score(m)
-  expect_equal(score, score2)
-
-  # Cox is better than baseline (Kaplan-Meier)
-  l2 = lrn("surv.coxph")
-  p2 = suppressWarnings(l2$train(t)$predict(t))
-  expect_lt(p2$score(m), KMscore)
-
-  # Another edge case: some dead rats and 1 only censored
-  p3 = p2$clone()$filter(row_ids = c(event_ids, cens_ids[1L]))
-  score = p3$score(m)
-  expect_numeric(score)
-  expect_true(score != p3$score(m2)) # since dead rats are removed
-  p3$data$distr = p3$distr
-  score2 = p3$score(m)
-  expect_equal(score, score2)
-
-  # Another edge case: some censored rats and 1 only dead
-  p4 = p2$clone()$filter(row_ids = c(cens_ids, event_ids[1L]))
-  score = p4$score(m)
-  expect_numeric(score)
-  expect_true(score != p4$score(m2)) # since the dead rat is removed
-  p4$data$distr = p4$distr
-  score2 = p4$score(m)
-  expect_equal(score, score2)
+  # ERV works as it should
+  m = msr("surv.logloss", ERV = TRUE)
+  expect_equal(unname(p_kaplan$score(m, task = t, train_set = t$row_ids)), 0)
+  expect_gt(unname(p_cox$score(m, task = t, train_set = t$row_ids)), 0)
 })
 
 test_that("distr measures work with 3d survival array", {