+++ tests/testthat/test_resample.R
@@ -1,8 +1,0 @@
-test_that("resample-step", {
-  task = mlr_tasks$get("rats")
-  learner = mlr_learners$get("surv.coxph")
-  resampling = rsmp("cv", folds = 3L)
-
-  rr = suppressWarnings(resample(task, learner, resampling))
-  expect_resample_result(rr)
-})